+++ freestanding-linux-example/src/main.rs
@@ -25,7 +25,7 @@ static mut HEAP_BITMAP: PageAligned<[u8; 2]> = heap_bitmap!(chunks = 16);
 #[global_allocator]
 // SAFETY: these statics are exclusively owned by `ALLOCATOR`.
 static ALLOCATOR: GlobalChunkAllocator<16> = unsafe {
-    GlobalChunkAllocator::<16>::new_raw(
+    GlobalChunkAllocator::<16>::new(
         core::ptr::slice_from_raw_parts_mut(
             core::ptr::addr_of_mut!(HEAP).cast(),
             256,